One-for-All-129 - Separated to the useSaveNewRecord.ts
