+++ apps/web/src/feature/record/list/operation/SaveNewRecordButton/index.tsx
@@ -1,11 +1,7 @@
 import type { Columns } from "@feature/app/schema/record-schema";
 import { notify } from "@components/ui/v4/notify/notify";
-import { generateId } from "@feature/app/function/generate-id";
-import { calcNextIndex } from "@feature/record/list/operation/SaveNewRecordButton/calc-next-index";
-import { useResetNewRecord } from "@feature/record/list/operation/useNewRecordState";
-import { useResetMode } from "@feature/record/list/useModeState";
-import { useAddRecord, useRecords } from "@feature/record/list/useRecordsState";
-import { useInsertRecordMutation } from "@v3/graphql/public";
+
+import { useSaveNewRecord } from "./useSaveNewRecord";
 
 export const SaveNewRecordButton = ({
   appId,
@@ -14,34 +10,11 @@ export const SaveNewRecordButton = ({
   appId: string;
   columns: Columns;
 }) => {
-  const { records } = useRecords();
-  const resetMode = useResetMode();
-  const resetNewRecord = useResetNewRecord();
-  const add = useAddRecord();
-
-  const [, mut] = useInsertRecordMutation();
+  const { saveNewRecord } = useSaveNewRecord(appId);
 
   const saveRecordHandler = async () => {
-    const index = calcNextIndex(records);
-    const recordId = generateId();
-
     try {
-      const { error } = await mut({
-        id: recordId,
-        appId,
-        index,
-        columns,
-      });
-      if (error) throw error;
-
-      add(index, {
-        recordId,
-        columns,
-        isEditing: false,
-      });
-
-      resetNewRecord();
-      resetMode();
+      await saveNewRecord(columns);
       notify("レコードを追加しました");
     } catch (e) {
       console.error(e);
